requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docx2pdf==0.1.8
easyocr==1.7.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.2
grpcio-status==1.62.2
httplib2==0.22.0
huggingface-hub==0.24.2
idna==3.7
imageio==2.34.2
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.5.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
setuptools==72.0.0
shapely==2.0.5
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tifffile==2024.7.24
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.70
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
